Extract lib-wx-effects
