+++ libraries/lib-wx-effects/EffectEditor.cpp
