+++ libraries/lib-wx-effects/EffectUIServices.cpp
