+++ libraries/lib-wx-effects/StatelessPerTrackEffect.cpp
