+++ libraries/lib-wx-effects/StatelessPerTrackEffect.h
