+++ src/effects/lv2/LV2Effect.h
@@ -20,7 +20,7 @@ class wxArrayString;
 #include "LV2Ports.h"
 #include "ShuttleGui.h"
 #include "SampleFormat.h"
-#include "../StatelessPerTrackEffect.h"
+#include "StatelessPerTrackEffect.h"
 
 // We use deprecated LV2 interfaces to remain compatible with older
 // plug-ins, so disable warnings
